ml-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
pydantic-settings==2.3.0
PyJWT==2.8.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis==5.0.4
sqlalchemy==2.0.30
river==0.21.0
joblib==1.4.2
shap==0.45.1
scikit-learn==1.5.0
httpx==0.27.0
celery==5.4.0
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pydantic[email]==2.7.2
slowapi==0.1.9
